pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "gaya"
version = "0.1.0"
description = "Simple data quality checks that just work."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
gaya = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gaya*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
